Fix stale states captured by lambdas in ViewModels
