+++ app/src/main/kotlin/com/hippo/ehviewer/client/data/FavListUrlBuilder.kt
@@ -29,6 +29,9 @@ data class FavListUrlBuilder(
     var prev: String? = null,
     var next: String? = null,
 ) {
+    val isLocal
+        get() = favCat == FAV_CAT_LOCAL
+
     fun setIndex(index: String?, isNext: Boolean) {
         next = index.takeIf { isNext }
         prev = index.takeUnless { isNext }
